+++ package/src/firestoreRules/collection.js
@@ -4,9 +4,9 @@
 * The pseudo-API (similar to Firebase or Firebase-admin APIs, though those are currently in flux).
 */
 
-import { getAs, setAs, deleteAs } from '../firestoreREST/index.js'
+import { getAs, patchAs, deleteAs } from '../firestoreREST/index.js'
 
-//function fail(msg) { throw new Error(msg); }
+function fail(msg) { throw new Error(msg); }
 
 /*
 * Unlike in the Firestore API, we allow authentication to be set after collection.
@@ -14,15 +14,25 @@ import { getAs, setAs, deleteAs } from '../firestoreREST/index.js'
 function collection(collectionPath) {   // string => { as: ... }
 
   return {
-    as: uid => ({   // (string|null) => firebase.firestore.CollectionReference -like
+    as: (o) => {   // ({ uid:string }|null) => firebase.firestore.CollectionReference -like
+      const uid = o ? o.uid : null;
 
-      doc: docPath => {
-        return documentAs(uid, `${collectionPath}/${docPath}`);
+      return {
+        doc: docPath => {
+          return documentAs(uid, `${collectionPath}/${docPath}`);
+        },
+
+        // Try to get any (all) documents of the collection.
+        //
+        // Note: With the JS clients, this returns a QueryDocumentSnapshot of documents in the collection. That does
+        //    not seem to be available in the REST API, and to keep things simple, we just check a single document.
+        //    Is this enough for the needs of checking access?
+        //
+        get: () => {    // () => Promise of true|string
+          return getAs(uid, "any");
+        }
       }
-
-      /*get: () => {
-      }*/
-    })
+    }
   };
 }
 
@@ -32,9 +42,13 @@ function documentAs(uid, documentPath) {   // (string, string) => { get, set, de
     get: () => {
       return getAs(uid, documentPath);
     },
+
+    // REST API has separate "patch" and "create". We just map "set" to the patching - is that ok?
+    //
     set: (data) => {
-      return setAs(uid, documentPath, data);
+      return patchAs(uid, documentPath, data);
     },
+
     delete: () => {
       return deleteAs(uid, documentPath)
     }