+++ package/src/firestoreRules/matchers.js
@@ -4,12 +4,6 @@
 * Conveniency functions for testing Firestore security rules.
 *
 * Imported automatically for the test Node.js environments (not global setup).
-*
-* Note: Used to have 'err instanceof FirebaseError' in the conditions, but that didn't work (maybe Jest interference
-*     stuff), but testing can as well be done with '.name'.
-*
-*     References:
-*       - firebase.FirebaseError -> https://firebase.google.com/docs/reference/js/firebase.FirebaseError
 */
 import { strict as assert } from 'assert'
 import { expect } from '@jest/globals'
@@ -17,43 +11,32 @@ import { expect } from '@jest/globals'
 import { PRIME_ROUND } from "./common.js"
 assert(!PRIME_ROUND);
 
-// Just importing enables them:
 expect.extend( {
   async toAllow(prom) {
-    try {
-      await prom;
+    const s = await prom;
+    if (s !== true) {
+      return { pass: false, message: () => myFmt(true, s) }
+    } else {
       return { pass: true };
-    } catch (err) {
-      if (err.name === 'FirebaseError' && err.code === 'permission-denied') {
-        return { pass: false, message: () => myFmt(true, err) }
-      } else {
-        return weird(err)
-      }
     }
   },
 
   async toDeny(prom) {
-    try {
-      await prom;
-      return { pass: false, message: () => myFmt(false) }   // there is no reason for passing
-    } catch (err) {
-      if (err.name === 'FirebaseError' && err.code === 'permission-denied') {
-        return { pass: true }
-      } else {
-        return weird(err)
-      }
+    const s = await prom;
+    if (s === true) {
+      return { pass: false, message: () => myFmt(false) }   // no reason for being allowed
+    } else {
+      return { pass: true };
     }
   }
 });
 
 function myFmt(expectedAllowed,reason) {   // (boolean, string | undefined) => string
-  const [a,b] = expectedAllowed ? ['allowed','denied'] : ['denied','allowed'];
+  const lookup = { false: "denied", true: "allowed" };
+  const a = lookup[expectedAllowed];
+  const b = lookup[!expectedAllowed];
 
   return `Expected ${a} but the Firebase operation was ${b.toUpperCase()}.` + reason ? ` [${reason}]`:'';
-}
-
-function weird(err) {   // assert failed within the code; not allow/deny
-  return { pass: false, message: () => err }    // e.g. "ReferenceError: xxx is not defined"
 }
 
 export { }